requirements.txt
numpy==1.18.0
pandas==1.2.3
docx2txt==0.8
regex==2021.3.17
beautifulsoup4==4.9.3
unicode==2.8
unidecode==1.2.0
requests==2.22.0
ocrmypdf==12.0.2
pluggy==0.13.1
textract
pdfplumber==0.5.24
nltk==3.5
matplotlib==3.2.2
gensim==4.0.1
plotly==4.14.3
pycryptodome==3.0.0
seaborn==0.11.1
pdfkit==0.6.1
wordcloud==1.8.1
Collections-extended==1.0.3
tet==0.3.2
PyPDF2==1.26.0
fitz==0.0.1.dev2
spacy==3.0.6
PyMuPDF==1.18.13
pdf2docx==0.5.1
docx==0.2.4
htmldocx==0.0.5
python-magic-bin==0.4.14
docx2pdf==0.1.7
scikit-learn==0.24.2
PDFNetPython3==9.0.0
stamper==0.1.1
tqdm==4.60.0
ipywidgets==7.6.3